Deduplicate `copyFdRange` with new `readOffset`
